x86/perf: Fix incorrect return type for perf_misc_flags
The return value of perf_misc_flags is assigned to the `misc`
field of the perf_event_header structure, and its type should
be `unsigned short`.

Fixes: 39447b3 ("perf: Enhance perf to allow for guest statistic
collection from host")

+++ arch/x86/events/core.c
@@ -2948,10 +2948,10 @@ unsigned long perf_instruction_pointer(struct pt_regs *regs)
 	return regs->ip + code_segment_base(regs);
 }
 
-unsigned long perf_misc_flags(struct pt_regs *regs)
+unsigned short perf_misc_flags(struct pt_regs *regs)
 {
 	unsigned int guest_state = perf_guest_state();
-	int misc = 0;
+	unsigned short misc = 0;
 
 	if (guest_state) {
 		if (guest_state & PERF_GUEST_USER)
