+++ arch/x86/include/asm/perf_event.h
@@ -537,7 +537,7 @@ struct x86_perf_regs {
 };
 
 extern unsigned long perf_instruction_pointer(struct pt_regs *regs);
-extern unsigned long perf_misc_flags(struct pt_regs *regs);
+extern unsigned short perf_misc_flags(struct pt_regs *regs);
 #define perf_misc_flags(regs)	perf_misc_flags(regs)
 
 #include <asm/stacktrace.h>
